move to standalone module, tsify
