+++ catalog/app/containers/Bucket/Bucket.js
@@ -1,21 +1,18 @@
 import * as R from 'ramda'
 import * as React from 'react'
-import * as redux from 'react-redux'
 import { Route, Switch, matchPath } from 'react-router-dom'
 import * as M from '@material-ui/core'
 
-import Error from 'components/Error'
 import Layout from 'components/Layout'
 import Placeholder from 'components/Placeholder'
-import * as authSelectors from 'containers/Auth/selectors'
-import { ThrowNotFound, createNotFound } from 'containers/NotFoundPage'
+import { ThrowNotFound } from 'containers/NotFoundPage'
 import { useBucketExistence } from 'utils/BucketCache'
 import * as NamedRoutes from 'utils/NamedRoutes'
 import * as BucketPreferences from 'utils/BucketPreferences'
 import * as RT from 'utils/reactTools'
 
 import BucketNav from './BucketNav'
-// import { ThrowNotFound, createNotFound } from './NotFoundPage'
+import CatchNotFound from './CatchNotFound'
 import { displayError } from './errors'
 
 const mkLazy = (load) =>
@@ -89,18 +86,6 @@ function BucketLayout({ bucket, section = false, children }) {
     />
   )
 }
-
-const CatchNotFound = createNotFound(() => {
-  const username = redux.useSelector(authSelectors.username)
-  const helpText = username
-    ? 'Try to navigate using one of these tabs above'
-    : 'Do you need to log in?'
-  return (
-    <M.Box mt={4}>
-      <Error headline="Nothing here" detail={helpText} />
-    </M.Box>
-  )
-})
 
 export default function Bucket({
   location,
